feat(nvim): use cmdheight=0, signcolumn=no on neovim
Neovim specific support for cmdheight=0,
so reclaiming that additional row.
:Messages can be used to view the last message,
which can otherwise be hidden as a result of the cmdline
disappearing. Also in general shows a history
of things that may not have shown in :messages.

The way nvim-lsp + nvim-cmp shows errors makes it
so signcolumn is not really needed,
so taking advantage of that to reclaim that column.

`laststatus=0` is another change, but it seems that
lualine is setting this value to show. Would preferbly
keep lualine to display if I set laststatus=3,
but that will require more work, so it is left as is for
the time being.