fix: replace rust_analyzer nvim-lspconfig with rustaceanvim
At this moment, this is just for better integration between
Cargo.toml and the editor. The default settings does not seem to
refresh suggestions when adding something new to Cargo.toml.

The expand macro functionality is also extremely nice and probably worth
using just by itself.

Should look into usage of DAP in the future too.