fix: update navigator-toolbox auto-hide style to work on Firefox@133
If `#urlbar` or `#urlbar-container` does not render visibly for Firefox,
it appears that Firefox fails to start at all. This is thus a hack
to keep it rendered offscreen.

The z-index that the sidebar was previous updated to made it overlap
over the navbar, which hides buttons and the url. That is thus lowered
to leave the navbar on top.

The position and z-index has a value in Firefox@133, so we force our
styling there with `!important`.