feat(river): add swaylock
swaylock was smaller than waylock,
sticking with it until I find a need for waylock instead.

swaylock seems to not crash after `systemctl suspend` and it mentions
that it implements the waylock protocol for locking, so I assume
that any concerns around this issue is resolved.

Uncertain if hardware key unlock is supported whenever I move to using
that as login.

No inclusion of swayidle for the time being; sticking to manual locking
for now.