chore: remove yambar TODO comment
This isn't important enough for me to swap to at the moment.
The way I use waybar, it's almost never on screen anyways,
so the power efficiency benefits are small.

Having this commit in the history is enough to remind me why I
wanted to potentially look into this in the future anyways.