fix: add qt6-wayland
Inexplicably, this fixes OBS being 1 FPS for xwayland (e.g. games).
This also breaks OBS tiling properly, at least on wideriver.
It causes overlap if the window is too small with the window next to it.
At the very least, recording is a more important functionality than
the visual tiling until a more proper solution could be found.