cfg(librewolf): remove auto-hide sidebar behavior
Some buggiest remains with doing so that ends up being a bit annoying.
Retains the appear on top behavior, but now can only be viewed on toggle
with always taking up 50% of the width.

Appear on top behavior is kept as having the page shift when toggled on
is annoying.