cfg(librewolf): remove auto-hide address bar behavior
This was a particularly buggy behavior, especially when
permissions are requested as I cannot interact with the
prompt that pops up half the time.

A toggle behavior seems potentially possible with the addition
of another extension such as `Userchrome Toggle`, but I do not
want to add another dependency for this.

Showing the address bar is probably overall safer for browsing
anyways, so removing this behavior entirely.

I'd like to have reduce it's size with the `zoom` css, but that
seems to cause the urlbar to not calculate it's width properly.
It could be worth revisiting in the future if how that is calculated
is changed.