feat(nvim): add initial config for blink.cmp
Some parts of the config is not functional or not
as functional as the nvim-cmp config,
but this does enable it on a basic level.

It seems like blink.setup has to be called
to allow the pop up menu to show up properly
after overriding the default config value
(or for another reason).

Snippet functionality in particular seems to be broken;
unsure how to fix that at the moment.