cfg(nvim): disable noice for cmdline and messages
I keep looking at the wrong part of the screen,
and it ended up not feeling like it would be worth getting
used to the top right location as it is not compatible with
a more typical configless vim setup.

It looks like the notify window can be configured,
but it's not a particular concern or desire for me at the moment.